# QgsFeature python optimisations