--- a/python/PyQt6/core/auto_generated/qgsattributes.sip.in
+++ b/python/PyQt6/core/auto_generated/qgsattributes.sip.in
@@ -88,19 +88,19 @@ typedef QVector<QVariant> QgsAttributes;
     }
     else if ( PyBool_Check( obj ) )
     {
-      qv->append( QVariant( PyObject_IsTrue( obj ) == 1 ) );
+      *outData++ = QVariant( PyObject_IsTrue( obj ) == 1 );
     }
     else if ( PyLong_Check( obj ) )
     {
-      qv->append( QVariant( PyLong_AsLongLong( obj ) ) );
+      *outData++ = QVariant( PyLong_AsLongLong( obj ) );
     }
     else if ( PyFloat_Check( obj ) )
     {
-      qv->append( QVariant( PyFloat_AsDouble( obj ) ) );
+      *outData++ = QVariant( PyFloat_AsDouble( obj ) );
     }
     else if ( PyUnicode_Check( obj ) )
     {
-      qv->append( QVariant( QString::fromUtf8( PyUnicode_AsUTF8( obj ) ) ) );
+      *outData++ = QVariant( QString::fromUtf8( PyUnicode_AsUTF8( obj ) ) );
     }
     else
     {