--- a/python/PyQt6/core/auto_generated/qgsfeature.sip.in
+++ b/python/PyQt6/core/auto_generated/qgsfeature.sip.in
@@ -34,14 +34,14 @@ geometry and a list of field/values attributes.
 
   public:
 
-    SIP_PYOBJECT __iter__();
+    SIP_PYOBJECT __iter__() /HoldGIL/;
 %MethodCode
     QgsAttributes attributes = sipCpp->attributes();
     PyObject *attrs = sipConvertFromType( &attributes, sipType_QgsAttributes, Py_None );
     sipRes = PyObject_GetIter( attrs );
 %End
 
-    SIP_PYOBJECT __getitem__( int key );
+    SIP_PYOBJECT __getitem__( int key ) /HoldGIL/;
 %MethodCode
     QgsAttributes attrs = sipCpp->attributes();
     if ( a0 < 0 || a0 >= attrs.count() )
@@ -56,7 +56,7 @@ geometry and a list of field/values attributes.
     }
 %End
 
-    SIP_PYOBJECT __getitem__( const QString &name );
+    SIP_PYOBJECT __getitem__( const QString &name ) /HoldGIL/;
 %MethodCode
     int fieldIdx = sipCpp->fieldNameIndex( *a0 );
     if ( fieldIdx == -1 )
@@ -71,7 +71,7 @@ geometry and a list of field/values attributes.
     }
 %End
 
-    void __setitem__( int key, SIP_PYOBJECT value /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ );
+    void __setitem__( int key, SIP_PYOBJECT value /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ ) /HoldGIL/;
 %MethodCode
     bool rv;
 
@@ -121,7 +121,7 @@ geometry and a list of field/values attributes.
     }
 %End
 
-    void __setitem__( const QString &key, SIP_PYOBJECT value /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ );
+    void __setitem__( const QString &key, SIP_PYOBJECT value /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ ) /HoldGIL/;
 %MethodCode
     int fieldIdx = sipCpp->fieldNameIndex( *a0 );
     if ( fieldIdx == -1 )
@@ -168,7 +168,7 @@ geometry and a list of field/values attributes.
     }
 %End
 
-    void __delitem__( int key );
+    void __delitem__( int key ) /HoldGIL/;
 %MethodCode
     if ( a0 >= 0 && a0 < sipCpp->attributes().count() )
       sipCpp->deleteAttribute( a0 );
@@ -179,7 +179,7 @@ geometry and a list of field/values attributes.
     }
 %End
 
-    void __delitem__( const QString &name );
+    void __delitem__( const QString &name ) /HoldGIL/;
 %MethodCode
     int fieldIdx = sipCpp->fieldNameIndex( *a0 );
     if ( fieldIdx == -1 )
@@ -191,19 +191,19 @@ geometry and a list of field/values attributes.
       sipCpp->deleteAttribute( fieldIdx );
 %End
 
-    Py_hash_t __hash__() const;
+    Py_hash_t __hash__() const /HoldGIL/;
 %MethodCode
     sipRes = qHash( *sipCpp );
 %End
 
-    QgsFeature( qint64 id = FID_NULL );
+    QgsFeature( qint64 id = FID_NULL ) /HoldGIL/;
 %Docstring
 Constructor for QgsFeature
 
 :param id: unique feature ID
 %End
 
-    QgsFeature( const QgsFields &fields, qint64 id = FID_NULL );
+    QgsFeature( const QgsFields &fields, qint64 id = FID_NULL ) /HoldGIL/;
 %Docstring
 Constructor for QgsFeature
 
@@ -211,26 +211,26 @@ Constructor for QgsFeature
 :param id: unique feature ID
 %End
 
-    QgsFeature( const QgsFeature &rhs );
+    QgsFeature( const QgsFeature &rhs ) /HoldGIL/;
 %Docstring
 Copy constructor
 %End
 
 
-    bool operator==( const QgsFeature &other ) const;
+    bool operator==( const QgsFeature &other ) const /HoldGIL/;
 
-    bool operator!=( const QgsFeature &other ) const;
+    bool operator!=( const QgsFeature &other ) const /HoldGIL/;
 
     virtual ~QgsFeature();
 
-    QgsFeatureId id() const;
+    QgsFeatureId id() const /HoldGIL/;
 %Docstring
 Returns the feature ID for this feature.
 
 .. seealso:: :py:func:`setId`
 %End
 
-    void setId( QgsFeatureId id );
+    void setId( QgsFeatureId id ) /HoldGIL/;
 %Docstring
 Sets the feature ``id`` for this feature.
 
@@ -245,7 +245,7 @@ Sets the feature ``id`` for this feature.
    layer or data provider, and the results will be unpredictable
 %End
 
-    QgsAttributes attributes() const;
+    QgsAttributes attributes() const /HoldGIL/;
 %Docstring
 Returns the feature's attributes.
 
@@ -265,7 +265,7 @@ its attributes:
 %End
 
 
-    SIP_PYOBJECT attributeMap() const /TypeHint="Dict[str, Optional[object]]"/;
+    SIP_PYOBJECT attributeMap() const /HoldGIL,TypeHint="Dict[str, Optional[object]]"/;
 %Docstring
 Returns the feature's attributes as a map of field name to value.
 
@@ -303,14 +303,14 @@ Returns the feature's attributes as a map of field name to value.
     }
 %End
 
-    int attributeCount() const;
+    int attributeCount() const /HoldGIL/;
 %Docstring
 Returns the number of attributes attached to the feature.
 
 .. versionadded:: 3.18
 %End
 
-    void setAttributes( const QgsAttributes &attrs );
+    void setAttributes( const QgsAttributes &attrs ) /HoldGIL/;
 %Docstring
 Sets the feature's attributes.
 
@@ -333,7 +333,7 @@ The number of provided attributes need to exactly match the number of the featur
 %End
 
 
-    bool setAttribute( int field, SIP_PYOBJECT attr /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ );
+    bool setAttribute( int field, SIP_PYOBJECT attr /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ ) /HoldGIL/;
 %Docstring
 Sets an attribute's value by field index.
 
@@ -411,7 +411,7 @@ Alternatively, in Python it is possible to directly set a field's value via the 
     sipRes = rv;
 %End
 
-    void initAttributes( int fieldCount );
+    void initAttributes( int fieldCount ) /HoldGIL/;
 %Docstring
 Initialize this feature with the given number of fields.
 
@@ -422,7 +422,7 @@ Discards any previously set attribute data.
 .. seealso:: :py:func:`resizeAttributes`
 %End
 
-    void resizeAttributes( int fieldCount );
+    void resizeAttributes( int fieldCount ) /HoldGIL/;
 %Docstring
 Resizes the attributes attached to this feature to the given number of fields.
 
@@ -439,7 +439,7 @@ end of the existing attributes.
 .. versionadded:: 3.18
 %End
 
-    void padAttributes( int count );
+    void padAttributes( int count ) /HoldGIL/;
 %Docstring
 Resizes the attributes attached to this feature by appending the specified ``count`` of NULL values to the end of the existing attributes.
 
@@ -449,7 +449,7 @@ Resizes the attributes attached to this feature by appending the specified ``cou
 %End
 
 
-    void deleteAttribute( int field );
+    void deleteAttribute( int field ) /HoldGIL/;
 %Docstring
 Clear's an attribute's value by its index.
 
@@ -487,7 +487,7 @@ Alternatively, in Python it is possible to directly `del` an attribute via its i
     }
 %End
 
-    bool isValid() const;
+    bool isValid() const /HoldGIL/;
 %Docstring
 Returns the validity of this feature.
 
@@ -497,7 +497,7 @@ invalid or to indicate a null feature.
 .. seealso:: :py:func:`setValid`
 %End
 
-    void setValid( bool validity );
+    void setValid( bool validity ) /HoldGIL/;
 %Docstring
 Sets the validity of the feature.
 
@@ -506,14 +506,14 @@ Sets the validity of the feature.
 .. seealso:: :py:func:`isValid`
 %End
 
-    bool hasGeometry() const;
+    bool hasGeometry() const /HoldGIL/;
 %Docstring
 Returns ``True`` if the feature has an associated geometry.
 
 .. seealso:: :py:func:`geometry`
 %End
 
-    QgsGeometry geometry() const;
+    QgsGeometry geometry() const /HoldGIL/;
 %Docstring
 Returns the geometry associated with this feature. If the feature has no geometry,
 an empty :py:class:`QgsGeometry` object will be returned.
@@ -523,7 +523,7 @@ an empty :py:class:`QgsGeometry` object will be returned.
 .. seealso:: :py:func:`setGeometry`
 %End
 
-    void setGeometry( const QgsGeometry &geometry );
+    void setGeometry( const QgsGeometry &geometry ) /HoldGIL/;
 %Docstring
 Set the feature's geometry.
 
@@ -536,7 +536,7 @@ Calling this method will automatically set the feature as valid (see :py:func:`~
 .. seealso:: :py:func:`clearGeometry`
 %End
 
-    void setGeometry( QgsAbstractGeometry *geometry /Transfer/ );
+    void setGeometry( QgsAbstractGeometry *geometry /Transfer/ ) /HoldGIL/;
 %Docstring
 Set the feature's ``geometry``.
 
@@ -577,7 +577,7 @@ Example
     sipCpp->setGeometry( std::unique_ptr< QgsAbstractGeometry>( a0 ) );
 %End
 
-    void clearGeometry();
+    void clearGeometry() /HoldGIL/;
 %Docstring
 Removes any geometry associated with the feature.
 
@@ -586,7 +586,7 @@ Removes any geometry associated with the feature.
 .. seealso:: :py:func:`hasGeometry`
 %End
 
-    void setFields( const QgsFields &fields, bool initAttributes = true  );
+    void setFields( const QgsFields &fields, bool initAttributes = true  ) /HoldGIL/;
 %Docstring
 Assigns a field map with the feature to allow attribute access by attribute name.
 
@@ -596,7 +596,7 @@ Assigns a field map with the feature to allow attribute access by attribute name
 .. seealso:: :py:func:`fields`
 %End
 
-    QgsFields fields() const;
+    QgsFields fields() const /HoldGIL/;
 %Docstring
 Returns the field map associated with the feature.
 
@@ -604,7 +604,7 @@ Returns the field map associated with the feature.
 %End
 
 
-    void setAttribute( const QString &name, SIP_PYOBJECT value /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ );
+    void setAttribute( const QString &name, SIP_PYOBJECT value /TypeHint="Optional[Union[bool, int, float, str, QVariant]]"/ ) /HoldGIL/;
 %Docstring
 Insert a value into attribute, by field ``name``.
 
@@ -680,7 +680,7 @@ Alternatively, in Python it is possible to directly set a field's value via the 
 %End
 
 
-    bool deleteAttribute( const QString &name );
+    bool deleteAttribute( const QString &name ) /HoldGIL/;
 %Docstring
 Clear's an attribute's value by its field ``name``.
 
@@ -728,7 +728,7 @@ Alternatively, in Python it is possible to directly `del` an attribute via its n
 %End
 
 
-    SIP_PYOBJECT attribute( const QString &name ) const;
+    SIP_PYOBJECT attribute( const QString &name ) const /HoldGIL/;
 %Docstring
 Lookup attribute value by attribute ``name``.
 
@@ -772,7 +772,7 @@ Alternatively, in Python it is possible to directly retrieve a field's value via
 %End
 
 
-    SIP_PYOBJECT attribute( int fieldIdx ) const;
+    SIP_PYOBJECT attribute( int fieldIdx ) const /HoldGIL/;
 %Docstring
 Lookup attribute value from its index.
 
@@ -817,7 +817,7 @@ Alternatively, in Python it is possible to directly retrieve a field's value via
 
 
 
-    bool isUnsetValue( int fieldIdx ) const;
+    bool isUnsetValue( int fieldIdx ) const /HoldGIL/;
 %Docstring
 Returns ``True`` if the attribute at the specified index is an unset value.
 
@@ -841,14 +841,14 @@ Returns ``True`` if the attribute at the specified index is an unset value.
     }
 %End
 
-    const QgsSymbol *embeddedSymbol() const;
+    const QgsSymbol *embeddedSymbol() const /HoldGIL/;
 %Docstring
 Returns the feature's embedded symbology, or ``None`` if the feature has no embedded symbol.
 
 .. versionadded:: 3.20
 %End
 
-    void setEmbeddedSymbol( QgsSymbol *symbol /Transfer/ );
+    void setEmbeddedSymbol( QgsSymbol *symbol /Transfer/ ) /HoldGIL/;
 %Docstring
 Sets the feature's embedded ``symbol``.
 
@@ -857,7 +857,7 @@ Ownership of ``symbol`` is transferred to the feature.
 .. versionadded:: 3.20
 %End
 
-    int fieldNameIndex( const QString &fieldName ) const;
+    int fieldNameIndex( const QString &fieldName ) const /HoldGIL/;
 %Docstring
 Utility method to get attribute index from name.