--- a/python/PyQt6/core/auto_generated/qgsfield.sip.in
+++ b/python/PyQt6/core/auto_generated/qgsfield.sip.in
@@ -40,7 +40,7 @@ length, and if applicable, precision.
               int len = 0,
               int prec = 0,
               const QString &comment = QString(),
-              QVariant::Type subType = QVariant::Invalid );
+              QVariant::Type subType = QVariant::Invalid ) /HoldGIL/;
 %Docstring
 Constructor. Constructs a new QgsField object.
 
@@ -58,7 +58,7 @@ Constructor. Constructs a new QgsField object.
                 this to QVariant.Invalid.
 %End
 
-    QgsField( const QgsField &other );
+    QgsField( const QgsField &other ) /HoldGIL/;
 %Docstring
 Copy constructor
 %End
@@ -66,10 +66,10 @@ Copy constructor
 
     virtual ~QgsField();
 
-    bool operator==( const QgsField &other ) const;
-    bool operator!=( const QgsField &other ) const;
+    bool operator==( const QgsField &other ) const /HoldGIL/;
+    bool operator!=( const QgsField &other ) const /HoldGIL/;
 
-    QString name() const;
+    QString name() const /HoldGIL/;
 %Docstring
 Returns the name of the field.
 
@@ -78,7 +78,7 @@ Returns the name of the field.
 .. seealso:: :py:func:`displayName`
 %End
 
-    QString displayName() const;
+    QString displayName() const /HoldGIL/;
 %Docstring
 Returns the name to use when displaying this field. This will be the
 field alias if set, otherwise the field name.
@@ -88,7 +88,7 @@ field alias if set, otherwise the field name.
 .. seealso:: :py:func:`alias`
 %End
 
-    QString displayNameWithAlias() const;
+    QString displayNameWithAlias() const /HoldGIL/;
 %Docstring
 Returns the name to use when displaying this field and adds the alias in parenthesis if it is defined.
 
@@ -103,7 +103,7 @@ represents.
 .. versionadded:: 3.12
 %End
 
-    QString displayType( bool showConstraints = false ) const;
+    QString displayType( bool showConstraints = false ) const /HoldGIL/;
 %Docstring
 Returns the type to use when displaying this field, including the length and precision of the datatype if applicable.
 
@@ -116,7 +116,7 @@ This will be used when the full datatype with details has to displayed to the us
 .. versionadded:: 3.14
 %End
 
-    QString friendlyTypeString() const;
+    QString friendlyTypeString() const /HoldGIL/;
 %Docstring
 Returns a user friendly, translated representation of the field type.
 
@@ -130,19 +130,19 @@ length, precision or constraint information.
 .. versionadded:: 3.14
 %End
 
-    QVariant::Type type() const;
+    QVariant::Type type() const /HoldGIL/;
 %Docstring
 Gets variant type of the field as it will be retrieved from data source
 %End
 
-    QVariant::Type subType() const;
+    QVariant::Type subType() const /HoldGIL/;
 %Docstring
 If the field is a collection, gets its element's type.
 When all the elements don't need to have the same type, this returns
 QVariant.Invalid.
 %End
 
-    QString typeName() const;
+    QString typeName() const /HoldGIL/;
 %Docstring
 Gets the field type. Field types vary depending on the data source. Examples
 are char, int, double, blob, geometry, etc. The type is stored exactly as
@@ -151,26 +151,26 @@ the data store reports it, with no attempt to standardize the value.
 :return: QString containing the field type
 %End
 
-    int length() const;
+    int length() const /HoldGIL/;
 %Docstring
 Gets the length of the field.
 
 :return: int containing the length of the field
 %End
 
-    int precision() const;
+    int precision() const /HoldGIL/;
 %Docstring
 Gets the precision of the field. Not all field types have a related precision.
 
 :return: int containing the precision or zero if not applicable to the field type.
 %End
 
-    QString comment() const;
+    QString comment() const /HoldGIL/;
 %Docstring
 Returns the field comment
 %End
 
-    QMap< int, QVariant > metadata() const;
+    QMap< int, QVariant > metadata() const /HoldGIL/;
 %Docstring
 Returns the map of field metadata.
 
@@ -182,7 +182,7 @@ Map keys should match values from the :py:class:`Qgis`.FieldMetadataProperty enu
 %End
 
 
-    QVariant metadata( int property ) const;
+    QVariant metadata( int property ) const /HoldGIL/;
 %Docstring
 Returns a specific metadata ``property``.
 
@@ -191,7 +191,7 @@ Returns a specific metadata ``property``.
 .. versionadded:: 3.32
 %End
 
-    void setMetadata( const QMap< int, QVariant > metadata );
+    void setMetadata( const QMap< int, QVariant > metadata ) /HoldGIL/;
 %Docstring
 Sets the map of field ``metadata``.
 
@@ -203,7 +203,7 @@ Map keys should match values from the :py:class:`Qgis`.FieldMetadataProperty enu
 %End
 
 
-    void setMetadata( int property, const QVariant &value );
+    void setMetadata( int property, const QVariant &value ) /HoldGIL/;
 %Docstring
 Sets a metadata ``property`` to ``value``.
 
@@ -212,65 +212,65 @@ Sets a metadata ``property`` to ``value``.
 .. versionadded:: 3.32
 %End
 
-    bool isNumeric() const;
+    bool isNumeric() const /HoldGIL/;
 %Docstring
 Returns if this field is numeric. Any integer or floating point type
 will return ``True`` for this.
 %End
 
-    bool isDateOrTime() const;
+    bool isDateOrTime() const /HoldGIL/;
 %Docstring
 Returns if this field is a date and/or time type.
 
 .. versionadded:: 3.6
 %End
 
-    void setName( const QString &name );
+    void setName( const QString &name ) /HoldGIL/;
 %Docstring
 Set the field name.
 
 :param name: Name of the field
 %End
 
-    void setType( QVariant::Type type );
+    void setType( QVariant::Type type ) /HoldGIL/;
 %Docstring
 Set variant type.
 %End
 
-    void setSubType( QVariant::Type subType );
+    void setSubType( QVariant::Type subType ) /HoldGIL/;
 %Docstring
 If the field is a collection, set its element's type.
 When all the elements don't need to have the same type, set this to
 QVariant.Invalid.
 %End
 
-    void setTypeName( const QString &typeName );
+    void setTypeName( const QString &typeName ) /HoldGIL/;
 %Docstring
 Set the field type.
 
 :param typeName: Field type
 %End
 
-    void setLength( int len );
+    void setLength( int len ) /HoldGIL/;
 %Docstring
 Set the field length.
 
 :param len: Length of the field
 %End
 
-    void setPrecision( int precision );
+    void setPrecision( int precision ) /HoldGIL/;
 %Docstring
 Set the field precision.
 
 :param precision: Precision of the field
 %End
 
-    void setComment( const QString &comment );
+    void setComment( const QString &comment ) /HoldGIL/;
 %Docstring
 Set the field comment
 %End
 
-    QgsDefaultValue defaultValueDefinition() const;
+    QgsDefaultValue defaultValueDefinition() const /HoldGIL/;
 %Docstring
 Returns the expression used when calculating the default value for the field.
 
@@ -280,7 +280,7 @@ Returns the expression used when calculating the default value for the field.
 .. seealso:: :py:func:`setDefaultValueDefinition`
 %End
 
-    void setDefaultValueDefinition( const QgsDefaultValue &defaultValueDefinition );
+    void setDefaultValueDefinition( const QgsDefaultValue &defaultValueDefinition ) /HoldGIL/;
 %Docstring
 Sets an expression to use when calculating the default value for the field.
 
@@ -290,21 +290,21 @@ Sets an expression to use when calculating the default value for the field.
 .. seealso:: :py:func:`defaultValueDefinition`
 %End
 
-    const QgsFieldConstraints &constraints() const;
+    const QgsFieldConstraints &constraints() const /HoldGIL/;
 %Docstring
 Returns constraints which are present for the field.
 
 .. seealso:: :py:func:`setConstraints`
 %End
 
-    void setConstraints( const QgsFieldConstraints &constraints );
+    void setConstraints( const QgsFieldConstraints &constraints ) /HoldGIL/;
 %Docstring
 Sets constraints which are present for the field.
 
 .. seealso:: :py:func:`constraints`
 %End
 
-    QString alias() const;
+    QString alias() const /HoldGIL/;
 %Docstring
 Returns the alias for the field (the friendly displayed name of the field ),
 or an empty string if there is no alias.
@@ -312,7 +312,7 @@ or an empty string if there is no alias.
 .. seealso:: :py:func:`setAlias`
 %End
 
-    void setAlias( const QString &alias );
+    void setAlias( const QString &alias ) /HoldGIL/;
 %Docstring
 Sets the alias for the field (the friendly displayed name of the field ).
 
@@ -321,7 +321,7 @@ Sets the alias for the field (the friendly displayed name of the field ).
 .. seealso:: :py:func:`alias`
 %End
 
-    Qgis::FieldConfigurationFlags configurationFlags() const;
+    Qgis::FieldConfigurationFlags configurationFlags() const /HoldGIL/;
 %Docstring
 Returns the Flags for the field (searchable, …).
 
@@ -330,7 +330,7 @@ Returns the Flags for the field (searchable, …).
 .. versionadded:: 3.34
 %End
 
-    void setConfigurationFlags( Qgis::FieldConfigurationFlags flags );
+    void setConfigurationFlags( Qgis::FieldConfigurationFlags flags ) /HoldGIL/;
 %Docstring
 Sets the Flags for the field (searchable, …).
 
@@ -344,7 +344,7 @@ Sets the Flags for the field (searchable, …).
 Formats string for display
 %End
 
-    static QString readableConfigurationFlag( Qgis::FieldConfigurationFlag flag );
+    static QString readableConfigurationFlag( Qgis::FieldConfigurationFlag flag ) /HoldGIL/;
 %Docstring
 Returns the readable and translated value of the configuration flag
 
@@ -415,14 +415,14 @@ Converts the provided variant to a compatible format
 
     operator QVariant() const;
 
-    void setEditorWidgetSetup( const QgsEditorWidgetSetup &v );
+    void setEditorWidgetSetup( const QgsEditorWidgetSetup &v ) /HoldGIL/;
 %Docstring
 Set the editor widget setup for the field.
 
 :param v: The value to set
 %End
 
-    QgsEditorWidgetSetup editorWidgetSetup() const;
+    QgsEditorWidgetSetup editorWidgetSetup() const /HoldGIL/;
 %Docstring
 Gets the editor widget setup for the field.
 
@@ -432,7 +432,7 @@ by manual field configuration.
 :return: the value
 %End
 
-    void setReadOnly( bool readOnly );
+    void setReadOnly( bool readOnly ) /HoldGIL/;
 %Docstring
 Make field read-only if ``readOnly`` is set to true. This is the case for
 providers which support generated fields for instance.
@@ -440,7 +440,7 @@ providers which support generated fields for instance.
 .. versionadded:: 3.18
 %End
 
-    bool isReadOnly() const;
+    bool isReadOnly() const /HoldGIL/;
 %Docstring
 Returns ``True`` if this field is a read-only field. This is the case for
 providers which support generated fields for instance.
@@ -448,7 +448,7 @@ providers which support generated fields for instance.
 .. versionadded:: 3.18
 %End
 
-    Qgis::FieldDomainSplitPolicy splitPolicy() const;
+    Qgis::FieldDomainSplitPolicy splitPolicy() const /HoldGIL/;
 %Docstring
 Returns the field's split policy, which indicates how field values should
 be handled during a split operation.
@@ -458,7 +458,7 @@ be handled during a split operation.
 .. versionadded:: 3.30
 %End
 
-    void setSplitPolicy( Qgis::FieldDomainSplitPolicy policy );
+    void setSplitPolicy( Qgis::FieldDomainSplitPolicy policy ) /HoldGIL/;
 %Docstring
 Sets the field's split ``policy``, which indicates how field values should
 be handled during a split operation.
@@ -468,7 +468,7 @@ be handled during a split operation.
 .. versionadded:: 3.30
 %End
 
-    Qgis::FieldDuplicatePolicy duplicatePolicy() const;
+    Qgis::FieldDuplicatePolicy duplicatePolicy() const /HoldGIL/;
 %Docstring
 Returns the field's duplicate policy, which indicates how field values should
 be handled during a duplicate operation.
@@ -478,7 +478,7 @@ be handled during a duplicate operation.
 .. versionadded:: 3.38
 %End
 
-    void setDuplicatePolicy( Qgis::FieldDuplicatePolicy policy );
+    void setDuplicatePolicy( Qgis::FieldDuplicatePolicy policy ) /HoldGIL/;
 %Docstring
 Sets the field's duplicate ``policy``, which indicates how field values should
 be handled during a duplicate operation.