… python/PyQt6/core/auto_generated/qgsfields.sip.in
… python/core/auto_generated/qgsfeature.sip.in
… python/core/auto_generated/qgsfield.sip.in
… python/core/auto_generated/qgsfields.sip.in
… src/core/qgsattributes.h
… src/core/qgsfeature.h
… src/core/qgsfield.h
… src/core/qgsfields.h