Optimise some more frequently called methods for use in PyQGIS