@nyalldawson 2024-05-22
@ptitjano 

> Have you been able to test it and measure the improvement?

With a similar test to the loop from https://github.com/qgis/QGIS/pull/57470, calling 10.5 million times

Before: 
18.3s

After:
16.4s

With the optimisations from https://github.com/qgis/QGIS/pull/57470:

Before (#57470 alone):
12.8s

After (this + #57470):
11.7s


So no where near as dramatic as #57470, but still worthwhile!